text2pandas-query-pl/requirements.txt
llama-index
llama-index-experimental
llama-index-llms-groq
llama-index-llms-gemini
# llama-index-embeddings-huggingface
# llama-parse
arize-phoenix
pyvis
python-dotenv
google-generativeai